chore(home): move module to data and domain layer
