+++ app/src/main/java/com/example/alpacamusclemaintenance/viewmodel/HomeViewModel.kt
@@ -2,8 +2,8 @@ package com.example.alpacamusclemaintenance.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import com.example.alpacamusclemaintenance.repository.HomeRepository
-import com.example.alpacamusclemaintenance.vo.Home
+import com.example.alpacamusclemaintenance.data.home.HomeRepository
+import com.example.alpacamusclemaintenance.domain.home.Home
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
