+++ app/src/main/res/layout/fragment_home.xml
@@ -8,7 +8,7 @@
 
     <variable
       name="home"
-      type="com.example.alpacamusclemaintenance.vo.Home" />
+      type="com.example.alpacamusclemaintenance.domain.home.Home" />
   </data>
 
   <androidx.constraintlayout.widget.ConstraintLayout
